power_spread/requirements.txt
pandas>=2.1,<3
numpy>=1.26,<3
requests>=2.31
python-dotenv>=1.0
gridstatus>=0.36
statsmodels>=0.14
holidays>=0.50
scipy>=1.11
pyarrow>=15
